[STUDY-20210209] 학습내용 복습하기 세팅
