+++ src/App.js
@@ -5,229 +5,15 @@ import logo from "./logo.svg";
 import "./App.css";
 
 function App() {
-  //ES6 destructuring 문법
-  let [title, titleChange] = useState([
-    "남자 추천 성수동 맛집",
-    "남자 추천 강남 맛집",
-    "남자 추천 상왕십리 맛집",
-  ]);
 
-  let [modal, modalChange] = useState(false);
-  let [titleNumber, titleNumberChange] = useState(0);
-  let [contents, contentsChange] = useState('');
-
-  // react 반복문
-  function testUI() {
-    var array = [];
-
-    for (var i = 0; i < 3; i++) {
-      array.push(<div>안녕</div>)
-    }
-
-    return array
-  }
-
-  function openModal() {
-    if (modal == false) {
-      modalChange(true)
-    } else {
-      modalChange(false);
-    }
-  }
-
-  function titleChangeGender() {
-
-    // state는 복사본을 만들어서 수정하세요
-    // 그런데 state는 deep copy를 합시다
-    // const newArray = title; 이건 값 공유, reference data type 특징
-    // title[0] = 21321; 이렇게 하면 안된다.
-    // 이렇게 사용하면 복사해서 독자적으로 사용하는 것이 아니라 값공유임!
-
-    // deep copy
-    // 리액트의 대원칙: immutable data
-    const newArray = [...title];
-    // 1. 0번째 title값 변경
-    // newArray[0] = '여자 추천 성수동 맛집'
-    // 2. 응용 문제 정렬하기 성공!
-    newArray.sort();
-
-    titleChange(newArray)
-  }
-
-
-
-  let [likeCount, likeCountChange] = useState(0);
-  // likeCountChange(100); likeCountChange는 함수
-
-
-  // map()은 유사 반복문 입니다.
-  const array = [2, 3, 4];
-
-  const newArray = array.map(function (a) {
-    return a * 2;
-  })
-
-  console.log(newArray);
-
-  let posts = "강남 고기 맛집";
-  // let style = { color: "pink", fontSize: "30px" };
-
-  // UI 만드는법
-  // 1. UI와 관련된 중요 정보들을 state로 저장해놓고
-  // 2. state에 따라서 UI가 수정되게 만듬
 
   return (
     <div className="App">
-      <div className="black-nav">Blog</div>
-      <button
-        type="button"
-        onClick={titleChangeGender}
-      >
-        성별 필터 버튼
-      </button>
-
-      {
-        // map에서 i는 0,1,2 ... 
-        title.map((titleText, i) => {
-          return (
-            <div className="list" key={i}>
-              <h3
-                onClick={() => { titleNumberChange(i) }}
-              >
-                {titleText}
-
-                <button
-                  type="button"
-                  onClick={() => { likeCountChange(likeCount + 1) }}
-                >
-                  ❤️
-                </button>
-                {likeCount}
-              </h3>
-              <p>2월 17일 발행</p>
-              <hr />
-            </div>
-          )
-        })
-      }
-
-      <div className="publish">
-        <input
-          onChange={(e) => { contentsChange(e.target.value) }}
-        />
-        <button
-          type="button"
-          onClick={() => {
-            let arrayCopy = [...title];
-            arrayCopy.unshift(contents);
-            titleChange(arrayCopy)
-          }}
-        >
-          저장
-        </button>
-      </div>
-
-      <div className="publish">
-        <input
-          onChange={(e) => { contentsChange(e.target.value) }}
-        />
-        <button
-
-          onClick={() => {
-            let arrayChange = [...title];
-            arrayChange.unshift(contents);
-            titleChange(arrayChange);
-          }}
-        >
-          저장
-        </button>
-      </div>
-
-
-      {/* unshift 배열앞에 추가하는 문법 */}
-      {/* 
-      - input된 value 변수로 만들고
-      - 저장버튼을 누르면 그 변수를 어디에 추가해야함 
-      */}
-
-      {/* {contents}
-      <br />
-      <input
-        onChange={(e) => { contentsChange(e.target.value) }}
-      /> */}
-
-      {/* onChange, onInput 두개 있음 */}
-
-
-      {/* 
-      <button onClick={() => { titleNumberChange(0) }}>버튼1</button>
-      <button onClick={() => { titleNumberChange(1) }}>버튼2</button>
-      <button onClick={() => { titleNumberChange(2) }}>버튼3</button> 
-      */}
-
-      <button
-        onClick={() => { modalChange(!modal) }}
-      >
-        모달버튼2
-      </button>
-      {/* <Test></Test> */}
-
-      <Profile></Profile>
-
-      {/* 
-        javascript 사용하려면 사용하면 된다
-        if문은 인식 못한다
-
-        1 < 3 ? console.log('true') : console.log('false')
-
-        1 < 3 if문의 조건식
-        ? console.log('true') 실행할 코드
-        : console.log('false') else 실행할 코드
-
-        리액트의 관습 null
-      */}
-
-
-      {/* 
-      App은 부모 컴포넌트
-      Modal은 자식 컴포넌트
-    
-      App이 가진 state를 사용 가능하게 전송가능 
-
-      이 문법의 이름은 props
-      */}
-      {
-        modal === true
-          ? <Modal title={title} titleNumber={titleNumber}></Modal>
-          : null
-      }
+      블로글 발행 복습
 
     </div>
   );
 }
-
-function Modal(props) {
-  return (
-    <>
-      <div className="modal">
-        <h2>{props.title[props.titleNumber]}</h2>
-        <p>날짜</p>
-        <p>상세내용</p>
-      </div>
-    </>
-  )
-}
-
-function Test(props) {
-  return (
-    <>
-      <div class="test">
-        <h2>{props.title[props.titleNumber]}</h2>
-      </div>
-    </>
-  )
-}
-
 export default App;
 
 /*
